requirements.txt
## Core requirements
django~=1.10.6
djangorestframework~=3.6.2
djangorestframework-filters~=0.9.1
djangorestframework-jwt
django-allauth~=0.27.0
django-cors-headers~=1.2.2
django-csv-utils
django-enumfields
django-filter~=0.15.3
django-minimal-user==0.0.1
django-rest-auth~=0.9.0
django-rest-swagger~=2.0.6
django-revproxy
wagtail~=1.9.0

## Python extras
pytz
python-dateutil
Pillow
requests
geopy
faker
pygments
markdown
beautifulsoup4
